pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "atomize-mvp"
version = "0.1.0"
description = "Atomize MVP content library generator"
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2.6",
  "python-dotenv>=1.0",
  "openai>=1.12",
  "faster-whisper>=1.0",
  "python-docx>=1.1",
  "rich>=13.7",
  "playwright>=1.42",
  "pillow>=10.2",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
